Limit stage-packages to direct dependencies
Snapcraft's suggested `stage-packages` include recursive dependencies,
which means they're not really suitable for complete inclusion. For
example, `libcrystalhd3` is only available for amd64 and i386, which
means the snap fails to build on other architectures. Limit
`stage-packages` only to direct dependencies, which are more likely to
be available on all architectures.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>