Support sending SMS messages
Add a new (optional) `to_sms_emails` configuration key that, if set,
will result in a short (SMS-length) email without an attachment being
sent to that list of email addresses. The idea is that this would be an
email address that uses an SMS gateway that results in a text message.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>